audio: apply mixer paths for HDMI and SPDIF
After a recent change, mixer path changes would no longer
be applied for HDMI and SPDIF output.

In select_devices(), reset_mixer_state() is always called,
but if the output device is seen as invalid, an early
return is done before the mixer updates are applied by
update_mixer_state(). These output devices are valid but
need no routing, so remove the use of *_INVALID completely.
This ensures that the mixer is reverted to the initial
state.

Note that HDMI+speaker or any other dual output case except
speaker+headset/headphone are still unsupported.

Change-Id: I206ee4689cbc9afab58c92c22647d6b4becb8528