Adjust the min and max frame rate settings for camcorder apps
These settings are configured by the SoC vendors to make sure that
both Camera and Video Encoder components are capable of recording videos
using Camera source with the min and max frame rates. Not necessarily all
the frame rates in between have to be supported, but the end point values (
i.e., the min and max values) must be supported.

On Manta, the min and max frame rates that the Camera supports is 15fps and 30fps,
respectively, according to SLSI camera engineers. This patch simply corrects
the settings to reflect the capability.

Bug: 7246962
Change-Id: Icf4a9ebcf4f8c742a943a1a07a994df5bb60475d