audio: fix headset mic capture
The audio path configuration was derived from the playback device
and the capture audio source only. When playback is idle, the playback
device defaults to speaker which prevents selection of wired or SCO
headset mic.
The fix consists in taking the capture device into account
when selecting the audio path when playback is idle.

Bug 7275723.

Change-Id: I13d125eaed13aaae706c675cbda7d35efa066619